feat : axe무기 추가 구현

+++ Assets/Scripts/Data/ItemData/ItemData.cs
@@ -19,7 +19,7 @@ public class ItemData : ScriptableObject
     public float baseDamage;
     public int baseCount;
     public float[] damages;
-    public int[] count;
+    public float[] count;
 
     [Header("Weapon")]
     public GameObject project;
